pyproject.toml
[tool.poetry]
name = "radar"
version = "0.1.0"
description = ""
authors = ["Leo Guinan <leo.guinan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
chromadb = "^0.5.5"
langchain = "^0.3.0"
scikit-learn = "^1.5.2"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.1"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
